- add ActDrawableResizeY - add action to use item - add right click event support to buttons - add support for using items (such as potions) - implement life/mana orbs - implement set size function for image - refactor ActGameSetProperty - refactor drawable actions - refactor item class formulas - refactor itemClass actions - refactor switch condition to support using previously defined case's actions - remove vs2015 configuration (no longer works) - rename gamefiles to gamefilesd - update player stats code - update gamefiles - use items such as potions - add initial hellfire gamefiles - hellfire supports more classes, jogging in town

+++ src/Actions/ActCondition.h
@@ -54,7 +54,7 @@ private:
 	}
 
 public:
-	ActIfCondition(unsigned conditionHash16_,
+	ActIfCondition(uint16_t conditionHash16_,
 		const VarOrPredicate& param1_,
 		const VarOrPredicate& param2_,
 		const std::shared_ptr<Action>& then_,
